chore(remote-config): Refactor remote config error logic
