+++ test/unit/remote-config/remote-config-api-client.spec.ts
@@ -19,10 +19,8 @@
 import * as _ from 'lodash';
 import * as chai from 'chai';
 import * as sinon from 'sinon';
-import {
-  FirebaseRemoteConfigError,
-  RemoteConfigApiClient
-} from '../../../src/remote-config/remote-config-api-client-internal';
+import { RemoteConfigApiClient } from '../../../src/remote-config/remote-config-api-client-internal';
+import { FirebaseRemoteConfigError } from '../../../src/remote-config/error';
 import { HttpClient } from '../../../src/utils/api-request';
 import * as utils from '../utils';
 import * as mocks from '../../resources/mocks';
